# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Plurality API",
    "description": "Description of Plurality API",
    "version": "1.0.0"
  },
  "host": "app.plurality.local:443",
  "basePath": "/",
...
  "schemes": [
    "https"
  ],
  "paths": {
...
    "/oauth-instagram/": {
      "get": {
        "tags": [
          "OAuth-Instagram"
        ],
        "description": "",
        "parameters": [
          {
            "name": "sse_id",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
...
    "/oauth-instagram/event": {
      "post": {
        "tags": [
          "OAuth-Instagram"
        ],
        "description": "",
        "parameters": [
          {
            "name": "x-sse-id",
            "in": "header",
            "type": "string"
          },
          {
            "name": "x-token-id",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/oauth-instagram/info": {
      "get": {
        "tags": [
          "OAuth-Instagram"
        ],
        "description": "",
        "parameters": [
          {
            "name": "x-token-id",
            "in": "header",
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "408": {
            "description": "Request Timeout"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/oauth-fortnite/": {
      "get": {
        "tags": [
          "OAuth-Fortnite"
        ],
        "description": "",
        "parameters": [
          {
            "name": "sse_id",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
...
    "/oauth-fortnite/event": {
      "post": {
        "tags": [
          "OAuth-Fortnite"
        ],
        "description": "",
        "parameters": [
          {
            "name": "x-sse-id",
            "in": "header",
            "type": "string"
          },
          {
            "name": "x-token-id",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/oauth-fortnite/info": {
      "get": {
        "tags": [
          "OAuth-Fortnite"
        ],
        "description": "",
        "parameters": [
          {
            "name": "x-token-id",
            "in": "header",
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "408": {
            "description": "Request Timeout"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  }
}